requirements.txt
# V2 Feature Fusion Pipeline
torch>=2.2.0
transformers>=4.40.0
accelerate>=0.29.0
pillow>=10.0.0
imagehash>=4.3.0
numpy>=1.26.0
scikit-learn>=1.4.0
pandas>=2.2.0
tqdm>=4.66.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pytesseract>=0.3.10
lightgbm>=4.3.0
sentence-transformers>=2.7.0
huggingface_hub>=0.20.0
